Do not CPU_GPU-migrate jobs that do not support it
Formerly, balancer support was synonymous with CPU_GPU support, and the list was reduced to "eligible" job beforehand; therefore, there was no need to check. However, with the advent of multiple balancer capabilities, this has changed.

Jobs without CPU_GPU support would still be migrated, but with incorrect qalter calls. We fix this by checking for CPU_GPU support.