Check PENDINGSINCE for corruption
Previously, Parser.Add_Pending_Since would check if PENDINGSINCE is already present in the Context, and add it if it isn’t.
Now, we also check if it is a valid number (narrowing it down to current date or, say, a year earlier, is an option for the future). If not, the variable is removed from the Context.
The value in memory will still be corrupt, leading to an exception later, but this prevents only the given job from being processed. In the next Balancer run, the job will get a new, valid PENDINGSINCE.